+++ src/main/java/org/opentripplanner/geocoder/AlternatingGeocoder.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.geocoder;
 
-import com.vividsolutions.jts.geom.Envelope;
+import org.locationtech.jts.geom.Envelope;
 
 
 public class AlternatingGeocoder implements Geocoder {