+++ src/main/java/org/opentripplanner/updater/street_notes/WFSNotePollingGraphUpdater.java
@@ -3,7 +3,7 @@ package org.opentripplanner.updater.street_notes;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.SetMultimap;
-import com.vividsolutions.jts.geom.Geometry;
+import org.locationtech.jts.geom.Geometry;
 import org.geotools.data.FeatureSource;
 import org.geotools.data.Query;
 import org.geotools.data.wfs.WFSDataStore;