+++ docs/Changelog.md
@@ -8,6 +8,7 @@
 - Remove the coupling to OneBusAway GTFS within OTP's internal model by creating new classes replacing the external classes (#2494)
 - Allow OTP to search more service days for transit service (#2592)
 - Allow itineraries in response to be sorted by duration (#2593)
+- Add support for GTFS-flex services: flag stops, deviated-route service, and call-and-ride (#2603)
 
 ## 1.3 (2018-08-03)